# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.2.0
  description: 巡梦 Dream Journal API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /dreams:
    get:
      operationId: listDreams
      tags: [dreams]
      summary: List all dream records
      responses:
        "200":
          description: List of dreams
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Dream"
    post:
      operationId: createDream
      tags: [dreams]
      summary: Create a new dream record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DreamInput"
      responses:
        "201":
          description: Created dream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dream"
...
  /dreams/{id}:
    get:
      operationId: getDream
      tags: [dreams]
      summary: Get a dream record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dream record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dream"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateDream
      tags: [dreams]
      summary: Update a dream record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DreamUpdate"
      responses:
        "200":
          description: Updated dream
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dream"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: deleteDream
      tags: [dreams]
      summary: Delete a dream record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    Dream:
      type: object
      required:
        - id
        - title
        - content
        - mood
        - clarity
        - isRecurring
        - createdAt
      properties:
        id:
          type: string
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          enum: [calm, anxious, happy, fearful, confused, nostalgic, strange]
        clarity:
          type: string
          enum: [vague, moderate, vivid]
        isRecurring:
          type: boolean
        createdAt:
          type: string
        characterId:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        symbols:
          type: array
          items:
            type: string
        emotionAnalysis:
          type: ["string", "null"]
        possibleConnection:
          type: ["string", "null"]
        companionReply:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    DreamInput:
      type: object
      required:
        - title
        - content
        - mood
        - clarity
        - isRecurring
      properties:
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          enum: [calm, anxious, happy, fearful, confused, nostalgic, strange]
        clarity:
          type: string
          enum: [vague, moderate, vivid]
        isRecurring:
          type: boolean
        characterId:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        symbols:
          type: array
          items:
            type: string
        emotionAnalysis:
          type: ["string", "null"]
        possibleConnection:
          type: ["string", "null"]
        companionReply:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
    DreamUpdate:
      type: object
      properties:
        title:
          type: string
        content:
          type: string
        mood:
          type: string
          enum: [calm, anxious, happy, fearful, confused, nostalgic, strange]
        clarity:
          type: string
          enum: [vague, moderate, vivid]
        isRecurring:
          type: boolean
        characterId:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        symbols:
          type: array
          items:
            type: string
        emotionAnalysis:
          type: ["string", "null"]
        possibleConnection:
          type: ["string", "null"]
        companionReply:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
...